Move all read config logic to config.go
